FISH-6037: fixing issue when processing java:app scope
