+++ appserver/connectors/connectors-internal-api/src/main/java/com/sun/appserv/connectors/internal/api/ConnectorConstants.java
@@ -212,7 +212,7 @@ public interface ConnectorConstants extends ResourceConstants {
     /**
      * Reserved sub-context where concurrent objects are bound with generated names.
      */
-    public static String CONCURRENT_JNDINAME_PREFIX=RESERVE_PREFIX+"/concurrent/";
+    public static String CONCURRENT_JNDINAME_PREFIX="concurrent/";
 
     /**
      *  Reserved sub-context where connector resource objects are bound with
