+++ nucleus/resources/src/main/java/org/glassfish/resourcebase/resources/api/ResourceConstants.java
@@ -142,7 +142,6 @@ public interface ResourceConstants {
     public static String JMS_CONNECTION_FACTORY_DEFINITION_JNDINAME_PREFIX = "__jms_connection_factory_definition/";
     public static String JMS_DESTINATION_DEFINITION_JNDINAME_PREFIX = "__jms_destination_definition/";
     public static String ADMINISTERED_OBJECT_DEFINITION_JNDINAME_PREFIX="__administered_object_definition/";
-    public static String CONCURRENT_MANAGEDEXECUTOR_DEFINTION_JNDINAME_PREFIX = "__defaultManagedExecutorService/";
 
     public static final String JAVA_SCOPE_PREFIX = "java:";
     public static final String JAVA_APP_SCOPE_PREFIX = "java:app/";
